Handle .ipa files without embedded.mobileprovision

+++ index.js
@@ -12,27 +12,30 @@ var glob = require("glob");
 var output = new tmp.Dir();
 
 module.exports = function (file, callback){
+  var data = {};
+
   fs.createReadStream(file)
     .pipe(unzip.Extract({ path: output.path }))
     .on('close', function(){
       var path = glob.sync(output.path + '/Payload/*/')[0];
 
-      var metadata = plist.parse(fs.readFileSync(path + 'Info.plist', 'utf8'));
+      data.metadata = plist.parse(fs.readFileSync(path + 'Info.plist', 'utf8'));
+
+      if(!fs.existsSync(path + 'Provisioning.plist')){
+        return callback(null, data);
+      }
 
       exec('security cms -D -i embedded.mobileprovision > Provisioning.plist', { cwd: path }, function(error) {
         if(error){
-          throw error;
+          return callback(error);
         }
 
-        var provisioning = plist.parse(fs.readFileSync(path + 'Provisioning.plist', 'utf8'));
-        delete provisioning.DeveloperCertificates;
+        data.provisioning = plist.parse(fs.readFileSync(path + 'Provisioning.plist', 'utf8'));
+        delete data.provisioning.DeveloperCertificates;
 
         rimraf.sync(output.path);
 
-        return callback({
-          metadata: metadata,
-          provisioning: provisioning
-        });
+        return callback(null, data);
       });
     });
 };
